Move embed_graph to verb-less operation
