+++ graphrag/index/operations/embed_graph/strategies/node_2_vec.py
@@ -9,7 +9,7 @@ import networkx as nx
 
 from graphrag.index.graph.embedding import embed_nod2vec
 from graphrag.index.graph.utils import stable_largest_connected_component
-from graphrag.index.verbs.graph.embed.typing import NodeEmbeddings
+from graphrag.index.operations.embed_graph.typing import NodeEmbeddings
 
 
 def run(graph: nx.Graph, args: dict[str, Any]) -> NodeEmbeddings:
