+++ graphrag/index/operations/embed_graph/typing.py
