+++ graphrag/index/verbs/__init__.py
@@ -10,7 +10,6 @@ from .graph import (
     cluster_graph,
     create_community_reports,
     create_graph,
-    embed_graph,
     layout_graph,
     merge_graphs,
     unpack_graph,
@@ -30,7 +29,6 @@ __all__ = [
     "concat",
     "create_community_reports",
     "create_graph",
-    "embed_graph",
     "entity_extract",
     "extract_covariates",
     "genid",
