+++ graphrag/index/verbs/graph/__init__.py
@@ -6,7 +6,6 @@
 from .clustering import cluster_graph
 from .compute_edge_combined_degree import compute_edge_combined_degree
 from .create import DEFAULT_EDGE_ATTRIBUTES, DEFAULT_NODE_ATTRIBUTES, create_graph
-from .embed import embed_graph
 from .layout import layout_graph
 from .merge import merge_graphs
 from .report import (
@@ -25,7 +24,6 @@ __all__ = [
     "compute_edge_combined_degree",
     "create_community_reports",
     "create_graph",
-    "embed_graph",
     "layout_graph",
     "merge_graphs",
     "prepare_community_reports",
